requirements.txt
opencv-python
pytesseract
easyocr
numpy
pillow
ultralytics
pika
kafka-python-ng
six
confluent-kafka
python-dotenv